Improve authorization
A few improvements to authorization, as well as adding basic documentation of
this feature to README. Can now store passwords as MD5 hashes instead of only
plaintext (mirroring functionality of kdb+ itself). Also made the authorization
code more generic, and easier to extend to other types of HTTP authorization.

Also added an additional example function to the example.q script, and a link
to recent blog post about this library.